move to basicfs and readd windows only solution
